requirements.txt
python-bitcoinrpc==1.0
click==8.1.7
